Tracking Inventory for ANY item, also from animals. This includes carried items - bills count the item immediately after it's produced as the crafter hauls it away.
It uses thingDef.EverHaulable, which honestly is probably everything crafted, and minifiable things are checked.

A chunk of code finding colonist items is consolidated for all the inventory types (was getting bloated)

The option for this could probably just be "check colonists" instead of "apparel" / "weapons" / "inventory" but the setting name is already saved, sooo... that's another thing.