add option to fix counting of items carried by non-humans
This includes hauled by animals/mechs, loaded onto caravan animals
and Anomaly ghouls (presumably, I do not have Anomaly). A visible
case of this is having a bill for e.g. flak vests that is per colonist,
and loading a couple of them on horses (to be equipped only before reaching
destination). Until the caravan leaves the map, colonists will be still
counted, items will not, and so temporarily the bill will enable
crafting more.